# docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "version": "1.0.0",
      "title": "AquaTrack Application",
      "license": {
        "name": "AquaTrack Live Frontend Application",
        "url": " "
      },
      "description": "This is a documentation of AquaTrack Application from Team 6!"
    },
    "servers": [
      {
        "url": " "
      }
    ],
...
    "paths": {
      "/water/daily": {
        "get": {
          "tags": [
            "Water"
          ],
          "summary": "Get daily water",
          "operationId": "getDailyWater",
          "description": "Get list of daily waters with this endpoint",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Get daily water",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully found daily water!"
                      },
                      "data": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/water"
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/401"
            }
          }
        }
      },
      "/water/monthly": {
        "get": {
          "tags": [
            "Water"
          ],
          "summary": "Get monthly water",
          "operationId": "getMonthlyWater",
          "description": "Get list of monthly waters with this endpoint",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Get monthly water",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully found monthly water!"
                      },
                      "data": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/water"
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/401"
            }
          }
        }
      },
      "/water": {
        "post": {
          "tags": [
            "Water"
          ],
          "summary": "Create data of water portion",
          "operationId": "addWater",
          "description": "Create data of water portion with payload",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "volume",
                    "date"
                  ],
                  "properties": {
                    "volume": {
                      "description": "The amount of water drunk, ml",
                      "type": "number",
                      "example": 250
                    },
                    "date": {
                      "description": "Time of drinking a portion of water",
                      "type": "string",
                      "example": "2025-03-05T10:00:00.000+00:00"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Create data of water portion",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 201
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully created data of water portion!"
                      },
                      "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/water"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/401"
            }
          }
        }
      },
      "/water/{id}": {
        "patch": {
          "tags": [
            "Water"
          ],
          "summary": "Update data of water portion",
          "operationId": "updateWater",
          "description": "Update the data about the portion of water by id with payload",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string",
                "example": "65ca67e7ae7f10c88b598008"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "volume": {
                      "description": "The amount of water drunk, ml",
                      "type": "number",
                      "example": 250
                    },
                    "date": {
                      "description": "Time of drinking a portion of water",
                      "type": "string",
                      "example": "2025-03-05T10:00:00.000+00:00"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Patch data of water portion",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully patched the data of water portion!"
                      },
                      "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/water"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "$ref": "#/components/responses/404"
            }
          }
        },
        "delete": {
          "tags": [
            "Water"
          ],
          "summary": "Delete data of water portion",
          "operationId": "deleteWater",
          "description": "Delete the data about the portion of water",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string",
                "example": "65ca67e7ae7f10c88b598008"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successfully the data about the portion of water deleted!"
            },
            "404": {
              "$ref": "#/components/responses/404"
            }
          }
        }
      },
...
      "/users/refresh": {
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Refresh Users Session",
          "operationId": "refreshUsersSession",
          "description": "Refresh Users Session in the Application",
          "security": [],
          "responses": {
            "200": {
              "description": "Successfully refreshed a session!",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully refreshed a session!"
                      },
                      "data": {
                        "type": "object",
                        "required": [
                          "accessToken"
                        ],
                        "properties": {
                          "accessToken": {
                            "type": "string",
                            "example": "2zQaqOpj5WrfbR32j0UAh7WAtaIoaRYkYNVZKg6T"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 401
                      },
                      "message": {
                        "type": "string",
                        "example": "UnauthorizedError"
                      },
                      "data": {
                        "type": "object",
                        "required": [
                          "message"
                        ],
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Session token expired"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer"
        }
      },
      "schemas": {
        "water": {
          "type": "object",
          "required": [
            "_id",
            "volume",
            "date",
            "userId"
          ],
          "properties": {
            "_id": {
              "description": "Backend-generated unique identifier",
              "type": "string",
              "example": "67a1f598a6ed272da1c633fd"
            },
            "volume": {
              "description": "The amount of water drunk, ml",
              "type": "number",
              "example": 250
            },
            "date": {
              "description": "Time of drinking a portion of water",
              "type": "string",
              "example": "2025-03-05T10:00:00.000+00:00"
            },
            "userId": {
              "description": "Record owner id",
              "type": "string",
              "example": "67a1f598a6ed272da1c632df"
            }
          }
        }
      },
      "responses": {
        "401": {
          "description": "Unauthorized error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 401
                  },
                  "message": {
                    "type": "string",
                    "example": "UnauthorizedError"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Access token expired"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "404": {
          "description": "Not Found Error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 404
                  },
                  "message": {
                    "type": "string",
                    "example": "NotFoundError"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Data not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
...
      }
    }
  }